refactor: perserve legacy CourseWaffleFlag
Updated flag moved to the temporary module __future__;
Updated all the imports to use flag from the __future__;
