+++ lms/djangoapps/course_home_api/toggles.py
@@ -2,7 +2,7 @@
 Toggles for course home experience.
 """
 
-from openedx.core.djangoapps.waffle_utils import CourseWaffleFlag
+from openedx.core.djangoapps.waffle_utils.__future__ import FutureCourseWaffleFlag as CourseWaffleFlag
 
 WAFFLE_FLAG_NAMESPACE = 'course_home'
