+++ lms/djangoapps/grades/config/waffle.py
@@ -6,7 +6,7 @@ waffle switches for the Grades app.
 
 from edx_toggles.toggles import WaffleSwitch
 
-from openedx.core.djangoapps.waffle_utils import CourseWaffleFlag
+from openedx.core.djangoapps.waffle_utils.__future__ import FutureCourseWaffleFlag as CourseWaffleFlag
 
 # Namespace
 WAFFLE_NAMESPACE = 'grades'
