+++ lms/djangoapps/instructor_task/config/waffle.py
@@ -5,7 +5,7 @@ waffle switches for the instructor_task app.
 
 from edx_toggles.toggles import WaffleSwitch
 
-from openedx.core.djangoapps.waffle_utils import CourseWaffleFlag
+from openedx.core.djangoapps.waffle_utils.__future__ import FutureCourseWaffleFlag as CourseWaffleFlag
 
 
 WAFFLE_NAMESPACE = 'instructor_task'
