+++ openedx/features/effort_estimation/toggles.py
@@ -2,7 +2,7 @@
 Feature toggles used for effort estimation.
 """
 
-from openedx.core.djangoapps.waffle_utils import CourseWaffleFlag
+from openedx.core.djangoapps.waffle_utils.__future__ import FutureCourseWaffleFlag as CourseWaffleFlag
 
 
 WAFFLE_FLAG_NAMESPACE = 'effort_estimation'
